+++ redbox/tests/api/test_format.py
@@ -3,19 +3,7 @@ import json
 
 
 from redbox.models.file import ChunkCreatorType
-from redbox.api.format import format_mcp_tool_response, SensitiveValue, MCPResponseMetadata
-
-
-class TestSensitiveValue:
-    @pytest.mark.parametrize("value", ["abc123", {"token": "x"}, None, 42])
-    def test_stores_and_retrieves(self, value):
-        assert SensitiveValue(value).get() == value
-
-    @pytest.mark.parametrize("value", ["super_secret", "abc123"])
-    def test_redacts_in_repr_and_str(self, value):
-        sv = SensitiveValue(value)
-        assert value not in repr(sv)
-        assert value not in str(sv)
+from redbox.api.format import format_mcp_tool_response, MCPResponseMetadata
 
 
 class TestFormatMCPToolResponse:
